# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/cache/clear": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Очищает кэш новостей (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Очистить кэш",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.AdminClearCacheResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/sources": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает информацию о всех источниках новостей (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить информацию об источниках",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v1.AdminSourceResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает статистику работы системы (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить статистику системы",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.AdminStatsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Авторизация по API ключу и получение JWT токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Данные для авторизации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/news": {
            "get": {
                "description": "Возвращает последние новости, агрегированные из всех источников",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Получить список новостей",
                "parameters": [
                    {
                        "maximum": 1000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Количество новостей (по умолчанию 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.NewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/validate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Проверяет валидность JWT токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Проверить токен",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ValidateTokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.News": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This breakthrough announcement changes the landscape..."
                },
                "id": {
                    "type": "string",
                    "example": "tech_news_1640995200_1"
                },
                "published_at": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "source": {
                    "type": "string",
                    "example": "Tech News"
                },
                "title": {
                    "type": "string",
                    "example": "Breaking: New Go Version Released"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/news/go-release"
                }
            }
        },
        "v1.AdminClearCacheResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Cache cleared successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "v1.AdminSourceResponse": {
            "type": "object",
            "properties": {
                "avg_response": {
                    "type": "string",
                    "example": "1.2s"
                },
                "last_check": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Tech News"
                },
                "news_count": {
                    "type": "integer",
                    "example": 2500
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "url": {
                    "type": "string",
                    "example": "https://tech-news-api.herokuapp.com/api/news"
                }
            }
        },
        "v1.AdminStatsResponse": {
            "type": "object",
            "properties": {
                "cache_hit_ratio": {
                    "type": "number",
                    "example": 0.85
                },
                "goroutines": {
                    "type": "integer",
                    "example": 50
                },
                "memory_usage": {
                    "type": "string",
                    "example": "128MB"
                },
                "sources_active": {
                    "type": "integer",
                    "example": 2
                },
                "total_news": {
                    "type": "integer",
                    "example": 5000
                },
                "total_requests": {
                    "type": "integer",
                    "example": 1000
                },
                "uptime": {
                    "type": "string",
                    "example": "24h30m"
                }
            }
        },
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "version": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
...
        "v1.LoginRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string",
                    "example": "your-api-key"
                }
            }
        },
        "v1.LoginResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "string",
                    "example": "24h"
                },
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "v1.NewsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 10
                },
                "news": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.News"
                    }
                },
                "version": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "v1.ValidateTokenResponse": {
            "type": "object",
            "properties": {
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "read",
                        "write"
                    ]
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}